Added support for overriding dataType via ApiModelProperty annotations
The only stipulation, in order to get this to work is that the dataType
is a package qualified class name.

Granted that arrays and generic types will not be available for this kind
of override. To guard from that situation if the type is only replacable
if the replacement is more specific than the already specified value.

fixes #602 and also addresses #554