Fixed AllowableValues on Implicit Parameters
There is still a case where the rendered output will not match the expectations if
the ApiImplicitParam is trying to enrich the properties of a property that is already
infered via springs machinery.

For e.g. any property that is infered via a @ModelAttribute annotation will already do the right thing
an infer the models before the @ApiImplicitPArameter values are overlayed. This is especially
fraught with problems if the underlying model property is a list of scalars rather than just a plain scalar
value... i.e. List<Enum>, List<String etc.

fixes #902