Fixed the failing test
The reason the test was failing was because the ModelAttributeParameterExpanderSpec in spring-web was trying to access a plugin that was in the swagger module. This creates an incorrect dependency link.

swagger-modules --- depend on --> spring modules.

This fix creates a new spec in the swagger module and removes the test
in the spring-web module and everything is right in the world :)