Added ability to override the base url
Since springfox relies on the position of swagger-ui.html in the URL, when that is not accurate in certain cases (#1615) or may be missing (#1707) we provide a prompt so the user can update the base url manually. This makes the UI useable even in cases where the api metadata is not available at the expected location.

fixes #1615
fixes #1707