Fixed an issue with render ApiImplicitParams
When the api implicit params does not indicate the type the parameter is rendered as (cookie, header, query etc.) it treats it as a body type. This fixes to not update the parameter type if its blank.

In addition this example demonstrates how to fix the reported issue via a contract test.

fixes #2029