Fixed the NPE when rendering complex path variable
When we have a complex path variable that is ignored via @ApiIgnore. Then the parameter type resolution falls back to using the ApiImplicitParam annotation value for parameter type (query, form etc.) As a result when its empty it ends up with a null value. The right thing to happen in this case is for the developer to fill in the type correctly to fix the problem.

This address that case of the NPE but doesnt try to provide a clever solution. Thus the example will render the operation to indicate that the bar parameter is a body parameter.

resolves #2148