Moved to use getters as a primary source for properties
This is a departure from the initial assumption that bean properties with backing fields will only be supported. This causes types like Pageable which are interfaces to no longer be supported. With this change we use getters as the primary source of truth and optionally for backwards compatibility we also use the fields, if present as a source for annotations

(1937)