Modified the builders to make it conducive to making changes given the uses
These builders are implementations of nested builders now. While they do the job, they need some work to make it intuitve. This leads to a bunch of yields in the builder and based on formatting its not really easy to dicern what the yield is returning. This needs to be reworked

(2022)