Added support for effective model
The idea behind effective models is to capture whether a model actually used the discriminators or not. For e.g. if a model is marked with a view discriminator and the dependent model does not have any effect of the discriminator as in the model is isomorphic with respect to the view discriminator then we remove the view from the effective model key. This is useful, in this example, to distinguish when naming a model if it should have `View` as a prefix or not in this case.

(2022)