Aligned dependencies
Switched to using the java-library plugin to better align the libraries
that are exposed as transitive dependencies. This is so that we can
deterministically say if an app is a SERVLET app or a REACTIVE app based
on the dependencies it brings along.

Also made the libraries declare beans that are mutually exclusive to the
paradigm.

(2022)